Release - Sprint 2

+++ travel-journal-api/src/main/java/travel/journal/api/service/TravelServiceImpl.java
@@ -8,7 +8,9 @@ import travel.journal.api.dto.travelJournal.inbound.TravelJournalDTO;
 import travel.journal.api.dto.travelJournal.outbound.TravelJournalDetailsDTO;
 import travel.journal.api.entities.Files;
 import travel.journal.api.entities.TravelJournal;
+import travel.journal.api.entities.User;
 import travel.journal.api.exception.ResourceNotFoundException;
+import travel.journal.api.exception.UnauthorizedAccesException;
 import travel.journal.api.repositories.TravelJournalRepository;
 
 import java.io.IOException;
@@ -21,11 +23,13 @@ public class TravelServiceImpl implements TravelService {
 
     private final TravelJournalRepository travelRepository;
     private final FilesServiceImpl filesService;
+    private final UserService userService;
     private final ModelMapper modelMapper;
 
-    public TravelServiceImpl(TravelJournalRepository travelRepository, FilesServiceImpl filesService, ModelMapper modelMapper) {
+    public TravelServiceImpl(TravelJournalRepository travelRepository, FilesServiceImpl filesService, UserService userService, ModelMapper modelMapper) {
         this.travelRepository = travelRepository;
         this.filesService = filesService;
+        this.userService = userService;
         this.modelMapper = modelMapper;
     }
 
@@ -57,46 +61,70 @@ public class TravelServiceImpl implements TravelService {
 
     @Override
     public TravelJournalDetailsDTO modifyTravelJournal(Integer id, TravelJournalDTO travelJournalDTO, MultipartFile file) throws IOException {
+        Optional<User> user = userService.getCurrentUser();
+        User checkuser = null;
+        if (user.isPresent()) {
+            checkuser = user.get();
+        }
+        
         Optional<TravelJournal> existingTravelOptional = travelRepository.findById(id);
-
         if (existingTravelOptional.isPresent()) {
+            if (user.isPresent()) {
+                if (checkuser.equals(existingTravelOptional.get().getUser())) {
+                    TravelJournal existingTravel = existingTravelOptional.get();
 
-            TravelJournal existingTravel = existingTravelOptional.get();
+                    Files modifiedImage = filesService.modifyImage(existingTravel.getCoverPhoto().getFileId(), file);
 
-            Files modifiedImage = filesService.modifyImage(existingTravel.getTravelId(), file);
+                    existingTravel.setHasCoverPhoto(modifiedImage.getFileContent().length > 0);
+                    existingTravel.setCoverPhoto(modifiedImage);
+                    existingTravel.setBudget(travelJournalDTO.getBudget());
+                    existingTravel.setDescription(travelJournalDTO.getDescription());
+                    existingTravel.setLocation(travelJournalDTO.getLocation());
+                    existingTravel.setEndDate(travelJournalDTO.getEndDate());
+                    existingTravel.setStartDate(travelJournalDTO.getStartDate());
+                    existingTravel.setHasCoverPhoto(existingTravel.getHasCoverPhoto());
+                    existingTravel.setNotesList(existingTravel.getNotesList());
 
-            existingTravel.setHasCoverPhoto(modifiedImage.getFileContent().length > 0);
-            existingTravel.setCoverPhoto(modifiedImage);
-            existingTravel.setBudget(travelJournalDTO.getBudget());
-            existingTravel.setDescription(travelJournalDTO.getDescription());
-            existingTravel.setLocation(travelJournalDTO.getLocation());
-            existingTravel.setEndDate(travelJournalDTO.getEndDate());
-            existingTravel.setStartDate(travelJournalDTO.getStartDate());
-            existingTravel.setHasCoverPhoto(existingTravel.getHasCoverPhoto());
-            existingTravel.setNotesList(existingTravel.getNotesList());
+                    TravelJournal modifiedTravel = travelRepository.save(existingTravel);
 
-            TravelJournal modifiedTravel = travelRepository.save(existingTravel);
-
-            return modelMapper.map(modifiedTravel, TravelJournalDetailsDTO.class);
+                    return modelMapper.map(modifiedTravel, TravelJournalDetailsDTO.class);
+                } else{
+                    throw new UnauthorizedAccesException("Current user is not authorized to update this travel journal");
+                }
+            }
         } else {
             throw new ResourceNotFoundException("Travel with id: " + id + " does not exist");
         }
+        return null;
     }
 
 
     @Override
     public void deleteTravelJournal(Integer id) {
+        Optional<User> user = userService.getCurrentUser();
+        User checkuser = null;
+        if (user.isPresent()) {
+            checkuser = user.get();
+        }
         Optional<TravelJournal> travelOptional = travelRepository.findById(id);
         if (travelOptional.isPresent()) {
+            if (user.isPresent()) {
+                if (checkuser.equals(travelOptional.get().getUser())) {
+                    TravelJournal travelToDelete = travelOptional.get();
+                    filesService.deleteImage(travelToDelete.getCoverPhoto().getFileId());
+                    travelRepository.deleteById(id);
 
-            TravelJournal travelToDelete = travelOptional.get();
-            filesService.deleteImage(travelToDelete.getCoverPhoto().getFileId());
-            travelRepository.deleteById(id);
+                } else {
+                    throw new UnauthorizedAccesException("Current user is not authorized to delete this travel journal");
+                }
+            }
 
-        } else {
-            throw new ResourceNotFoundException("Travel with id: " + id + " does not exist");
         }
+            else{
+                    throw new ResourceNotFoundException("Travel with id: " + id + " does not exist");
+                }
     }
+
 
     @Override
     public List<TravelJournalDetailsDTO>getUserTravelJournal(int userId) {
